+++ ESP-sc-gway/loraModem.h
@@ -1,7 +1,7 @@
 // 1-channel LoRa Gateway for ESP8266
 // Copyright (c) 2016, 2017 Maarten Westenberg version for ESP8266
-// Version 4.0.7
-// Date: 2017-07-22
+// Version 4.0.8
+// Date: 2017-08-05
 //
 // 	based on work done by Thomas Telkamp for Raspberry PI 1ch gateway
 //	and many other contributors.
@@ -80,27 +80,32 @@ unsigned long hopTime=0;
 volatile bool inIntr = 0;
 #endif
 
-// Definition of the GPIO pins used by the Gateway
+#if _PIN_OUT==1
+// Definition of the GPIO pins used by the Gateway for Hallard type boards
 struct pins {
 	uint8_t dio0=15;	// GPIO15 / D8. For the Hallard board shared between DIO0/DIO1/DIO2
 	uint8_t dio1=15;	// GPIO15 / D8. Used for CAD, may or not be shared with DIO0
 	uint8_t dio2=15;	// GPIO15 / D8. Used for frequency hopping, don't care
 	uint8_t ss=16;		// GPIO16 / D0. Select pin connected to SPIO16 / D0
 	uint8_t rst=0;		// GPIO0 / D3. Reset pin not used	
-// MISO 12 / D6
-// MOSI 13 / D7
-// CLK 14 / D5
-// SS 16 / D0
+	// MISO 12 / D6
+	// MOSI 13 / D7
+	// CLK  14 / D5
+	// SS   16 / D0
 } pins;
-
+#elif _PIN_OUT==2
 // For ComResult gateway PCB use the following settings
-//struct pins {
-//	uint8_t dio0=5;		// GPIO5 / D1. For the Hallard board shared betweenDio0/DIO1/DIO2
-//	uint8_t dio1=4;		// GPIO4 / D2. Used for CAD, may or not be shared with DIO0
-//	uint8_t dio2=0;		// GPIO0 / D3. Used for frequency hopping, don't care
-//	uint8_t ss=15;		// GPIO15 / D8. Select pin connected to SPIO16
-//	uint8_t rst=0;		// Reset pin not used	
-//} pins;
+struct pins {
+	uint8_t dio0=5;		// GPIO5 / D1. Dio0 used for one frequency and one SF
+	uint8_t dio1=4;		// GPIO4 / D2. Used for CAD, may or not be shared with DIO0
+	uint8_t dio2=0;		// GPIO0 / D3. Used for frequency hopping, don't care
+	uint8_t ss=15;		// GPIO15 / D8. Select pin connected to SPIO16
+	uint8_t rst=0;		// GPIO0 / D3. Reset pin not used	
+} pins;
+#else
+	// Use your own pin definitions
+#error "Pin Definitions _PIN_OUT must be 1(HALLARD) or 2 (COMRESULT)"
+#endif
 
 // STATR contains the statictis that are kept by message. 
 // Ech time a message is received or sent the statistics are updated.